test: one more test to reach 83%
