+++ src/kili/adapters/kili_api_gateway/__init__.py
@@ -1,7 +1,7 @@
 """Kili API Gateway module for interacting with Kili."""
 
 from kili.adapters.http_client import HttpClient
-from kili.adapters.kili_api_gateway.api_key import ApiKeyOperationMixin
+from kili.adapters.kili_api_gateway.api_key.operations_mixin import ApiKeyOperationMixin
 from kili.adapters.kili_api_gateway.asset import AssetOperationMixin
 from kili.adapters.kili_api_gateway.cloud_storage import CloudStorageOperationMixin
 from kili.adapters.kili_api_gateway.issue import IssueOperationMixin
